# Optimize Material Update Pointers