--- a/.github/workflows/main.yml
+++ b/.github/workflows/main.yml
@@ -160,6 +160,7 @@ jobs:
       - run: python benchmark/benchmark_big_model.py  
         name: Benchmark against big model
 
+        
   changelog-test:
     runs-on: ubuntu-latest
     if: github.ref != 'refs/heads/main'