@MicahGale 2024-08-28
> Consider printing fewer stats, but looks good.

Did you look at the output? The number of row filters is applied before the function regex. So it's not printing 300 lines.